ReferenceFiniteElements.AbstractHex

ReferenceFiniteElements.AbstractQuad

ReferenceFiniteElements.AbstractTet

ReferenceFiniteElements.AbstractTri

ReferenceFiniteElements.Edge

ReferenceFiniteElements.Hex8

ReferenceFiniteElements.Interpolants

ReferenceFiniteElements.Quad4

ReferenceFiniteElements.Quad9

ReferenceFiniteElements.ReferenceFE

ReferenceFiniteElements.ReferenceFE

ReferenceFiniteElements.ReferenceFEType

ReferenceFiniteElements.SimplexTri

ReferenceFiniteElements.Tet10

ReferenceFiniteElements.Tet4

ReferenceFiniteElements.Tri3

ReferenceFiniteElements.Tri6

Base.show

ReferenceFiniteElements.degree

ReferenceFiniteElements.edges

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_stencil

ReferenceFiniteElements.element_type

ReferenceFiniteElements.float_type

ReferenceFiniteElements.int_type

ReferenceFiniteElements.num_dimensions

ReferenceFiniteElements.num_dimensions

ReferenceFiniteElements.num_nodes

ReferenceFiniteElements.num_nodes_per_element

ReferenceFiniteElements.num_q_points

ReferenceFiniteElements.num_q_points

ReferenceFiniteElements.quadrature_points

ReferenceFiniteElements.quadrature_points_and_weights

ReferenceFiniteElements.quadrature_points_and_weights

ReferenceFiniteElements.quadrature_weights

ReferenceFiniteElements.setup_interpolants_simplex_tri!

ReferenceFiniteElements.setup_quad_quadrature_weights!

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_gradients

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_hessians

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.shape_function_values

ReferenceFiniteElements.vertex_nodes

ReferenceFiniteElements.AbstractHex — Type

ReferenceFiniteElements.AbstractQuad — Type

ReferenceFiniteElements.AbstractTet — Type

ReferenceFiniteElements.AbstractTri — Type

ReferenceFiniteElements.Edge — Type

TODO fix the 1 below and make multiple edge types maybe?

ReferenceFiniteElements.Hex8 — Type

degree::Int64

ReferenceFiniteElements.Interpolants — Type

ξ::Any

w::Any

N::Any

∇N_ξ::Any

∇∇N_ξ::Any

Interpolant container for a single quadrature point

ReferenceFiniteElements.Quad4 — Type

degree::Int64

ReferenceFiniteElements.Quad9 — Type

degree::Int64

ReferenceFiniteElements.ReferenceFE — Type

ref_fe_type::ReferenceFiniteElements.ReferenceFEType

nodal_coordinates::AbstractVecOrMat

edge_nodes::AbstractMatrix

face_nodes::AbstractMatrix

interior_nodes::AbstractVector

interpolants::Any

ReferenceFE is the main type.

This type defines the nodal coordinates, face nodes, interior nodes and interpolants

ReferenceFiniteElements.ReferenceFE — Method

Constructor for ReferenceFE

ReferenceFiniteElements.ReferenceFEType — Type

Type to define new element shapes

ReferenceFiniteElements.SimplexTri — Type

n_nodes::Int64

degree::Int64

ReferenceFiniteElements.Tet10 — Type

degree::Int64

ReferenceFiniteElements.Tet4 — Type

degree::Int64

ReferenceFiniteElements.Tri3 — Type

degree::Int64

ReferenceFiniteElements.Tri6 — Type

degree::Int64

Base.show — Method

show(io::IO, e::ReferenceFE)

show(io, e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:50.

ReferenceFiniteElements.degree — Method

degree(e::ReferenceFiniteElements.ReferenceFEType) -> Any

Returns the quadrature degree of a ReferenceFEType

degree(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFETypes.jl:10.

ReferenceFiniteElements.edges — Method

edges(e::ReferenceFE)

Returns the edge list

edges(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:174.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Hex8,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Matrix, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Hex8.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Quad4,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Any, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad4.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Quad9,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Any, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad9.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    e::SimplexTri,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Any, Any, Any}

element_stencil(e, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/SimplexTri.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Tet10,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Matrix, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet10.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Tet4,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Matrix, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet4.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Tri3,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Any, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri3.jl:3.

ReferenceFiniteElements.element_stencil — Method

element_stencil(
    _::Tri6,
    _::Type{Itype<:Integer},
    _::Type{Ftype<:AbstractFloat}
) -> Tuple{Matrix, Matrix, Any, Vector}

element_stencil(_, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri6.jl:3.

ReferenceFiniteElements.element_type — Method

element_type(
    _::ReferenceFE{Itype, Ftype, N, D, Q, RefFE<:ReferenceFiniteElements.ReferenceFEType, S, VOM, M1, M2, V}
) -> Type{RefFE} where RefFE<:ReferenceFiniteElements.ReferenceFEType

Returns the element type

element_type(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:131.

ReferenceFiniteElements.float_type — Method

float_type(
    _::ReferenceFE{Itype, Ftype, N, D, Q, RefFE<:ReferenceFiniteElements.ReferenceFEType, S, VOM, M1, M2, V}
) -> Any

Returns the float type used to store nodal coordinates and interpolation arrays

float_type(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:153.

ReferenceFiniteElements.int_type — Method

int_type(
    _::ReferenceFE{Itype, Ftype, N, D, Q, RefFE<:ReferenceFiniteElements.ReferenceFEType, S, VOM, M1, M2, V}
) -> Any

Returns the integer type used to store node ids and such

int_type(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:146.

ReferenceFiniteElements.num_dimensions — Method

num_dimensions(e::ReferenceFE) -> Any

Returns number of dimensions

num_dimensions(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:159.

ReferenceFiniteElements.num_dimensions — Method

num_dimensions(
    _::ReferenceFiniteElements.ReferenceFEType{N, D, Q}
) -> Any

Returns the number of dimensions for a ReferenceFEType

num_dimensions(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFETypes.jl:20.

num_dimensions(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:159.

ReferenceFiniteElements.num_nodes — Method

num_nodes(
    _::ReferenceFiniteElements.ReferenceFEType{N, D, Q}
) -> Any

Returns the number of nodes for a ReferenceFEType

num_nodes(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFETypes.jl:15.

ReferenceFiniteElements.num_nodes_per_element — Method

num_nodes_per_element(e::ReferenceFE) -> Any

Returns number of nodes per element

num_nodes_per_element(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:164.

ReferenceFiniteElements.num_q_points — Method

num_q_points(e::ReferenceFE) -> Any

Returns number of quadrature points

num_q_points(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:169.

ReferenceFiniteElements.num_q_points — Method

num_q_points(
    _::ReferenceFiniteElements.ReferenceFEType{N, D, Q}
) -> Any

Returns the number of quadrature porints for a ReferenceFEType

num_q_points(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFETypes.jl:25.

num_q_points(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:169.

ReferenceFiniteElements.quadrature_points — Method

quadrature_points(e::ReferenceFE) -> Any

Returns all quadrature points

quadrature_points(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:76.

ReferenceFiniteElements.quadrature_points_and_weights — Method

quadrature_points_and_weights(
    e::ReferenceFiniteElements.AbstractHex,
    _::Type{A<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    _::Type{T<:Number}
) -> Tuple{Any, Any}

quadrature_points_and_weights(e, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/HexCommon.jl:18.

ReferenceFiniteElements.quadrature_points_and_weights — Method

quadrature_points_and_weights(
    e::ReferenceFiniteElements.AbstractQuad,
    _::Type{A<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    _::Type{T<:Number}
) -> Tuple{Any, Any}

quadrature_points_and_weights(e, _, _)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/QuadCommon.jl:41.

ReferenceFiniteElements.quadrature_weights — Method

quadrature_weights(e::ReferenceFE) -> Any

Returns all quadrature weights

quadrature_weights(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:87.

ReferenceFiniteElements.setup_interpolants_simplex_tri! — Method

setup_interpolants_simplex_tri!(
    Ns,
    ∇N_ξs,
    ∇∇N_ξs,
    e,
    ξs,
    coordinates
)

setup_interpolants_simplex_tri!(
    Ns,
    ∇N_ξs,
    ∇∇N_ξs,
    e,
    ξs,
    coordinates
)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/SimplexTri.jl:48.

ReferenceFiniteElements.setup_quad_quadrature_weights! — Method

setup_quad_quadrature_weights!(w_return, ws)

setup_quad_quadrature_weights!(w_return, ws)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/QuadCommon.jl:27.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(e::ReferenceFE, q::Int64) -> Any

Returns a specific quadrature point's shape function gradient indexed by q

shape_function_gradients(e, q)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:113.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(e::ReferenceFE) -> Any

Returns all shape function gradients

shape_function_gradients(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:108.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Hex8,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Hex8.jl:110.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Quad4,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad4.jl:34.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Quad9,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad9.jl:58.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Tet10,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet10.jl:111.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Tet4,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet4.jl:77.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Tri3,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri3.jl:33.

ReferenceFiniteElements.shape_function_gradients — Method

shape_function_gradients(
    e::Tri6,
    _::Type{A1<:Union{StaticArraysCore.MMatrix, StaticArraysCore.SMatrix}},
    ξ::AbstractArray{<:Number}
) -> Any

shape_function_gradients(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri6.jl:55.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(e::ReferenceFE, i::Int64) -> Any

Returns a specific quadrature point's shape function hessian indexed by q

shape_function_hessians(e, i)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:123.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(e::ReferenceFE) -> Any

Returns all shape function hessians

shape_function_hessians(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:118.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Hex8,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Hex8.jl:147.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Quad4,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad4.jl:53.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Quad9,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad9.jl:88.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Tet10,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet10.jl:157.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Tet4,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet4.jl:91.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Tri3,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri3.jl:44.

ReferenceFiniteElements.shape_function_hessians — Method

shape_function_hessians(
    e::Tri6,
    _::Type{A1<:Union{StaticArraysCore.MArray, StaticArraysCore.SArray}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_hessians(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri6.jl:78.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(e::ReferenceFE, q::Int64) -> Any

Returns a specific quadrature point's shape function value indexed by q

shape_function_values(e, q)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:103.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(e::ReferenceFE) -> Any

Returns all shape function values

shape_function_values(e)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:98.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Quad4,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad4.jl:21.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Quad9,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Quad9.jl:40.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Tet10,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet10.jl:88.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Tet4,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tet4.jl:64.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Tri3,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri3.jl:21.

ReferenceFiniteElements.shape_function_values — Method

shape_function_values(
    e::Tri6,
    _::Type{A1<:Union{StaticArraysCore.SVector, StaticArraysCore.MVector}},
    ξ::AbstractVector{<:Number}
) -> Any

shape_function_values(e, _, ξ)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/implementations/Tri6.jl:39.

ReferenceFiniteElements.vertex_nodes — Method

vertex_nodes(
    _::ReferenceFE{Itype, Ftype, N, D, Q, RefFE<:ReferenceFiniteElements.ReferenceFEType, S, VOM, M1, M2, V}
) -> Base.OneTo

Returns the nodes of vertices TODO this is probably not very useful

vertex_nodes(_)

defined at /home/runner/work/ReferenceFiniteElements.jl/ReferenceFiniteElements.jl/src/ReferenceFEs.jl:139.