Element types

Below is the currently implemented set of elements organized by vertices, edges, faces, and volumes respectively. If you would like to see an additional element type supported, please file and issue or open a PR.

0-Dimensional Elements

ReferenceFiniteElements.Vertex — Type

struct Vertex <: ReferenceFiniteElements.AbstractVertex

1-Dimensional Elements

ReferenceFiniteElements.Edge — Type

struct Edge{PT, PD} <: ReferenceFiniteElements.AbstractEdge{PT, PD}

2-Dimensional Elements

ReferenceFiniteElements.Quad — Type

struct Quad{PT, PD} <: ReferenceFiniteElements.AbstractQuad{PT, PD}

ReferenceFiniteElements.Tri — Type

struct Tri{PT, PD} <: ReferenceFiniteElements.AbstractTri{PT, PD}

3-Dimensional Elements

ReferenceFiniteElements.Hex — Type

struct Hex{PT, PD} <: ReferenceFiniteElements.AbstractHex{PT, PD}

ReferenceFiniteElements.Tet — Type

struct Tet{PT, PD} <: ReferenceFiniteElements.AbstractTet{PT, PD}

Polynomial types

ReferenceFiniteElements.Hermite — Type

struct Hermite <: ReferenceFiniteElements.AbstractPolynomialType

ReferenceFiniteElements.Lagrange — Type

struct Lagrange <: ReferenceFiniteElements.AbstractPolynomialType

ReferenceFiniteElements.NoInterpolation — Type

struct NoInterpolation <: ReferenceFiniteElements.AbstractPolynomialType

ReferenceFiniteElements.RaviartThomas — Type

struct RaviartThomas <: ReferenceFiniteElements.AbstractPolynomialType

ReferenceFiniteElements.Serendipity — Type

struct Serendipity <: ReferenceFiniteElements.AbstractPolynomialType

Quadrature types

ReferenceFiniteElements.GaussLegendre — Type

struct GaussLegendre <: AbstractQuadratureType

ReferenceFiniteElements.GaussLobattoLegendre — Type

struct GaussLobattoLegendre <: AbstractQuadratureType