Type hierarchy

The root of the type hierarchy tree in ReferenceFiniteElements is the following abstract type.

ReferenceFiniteElements.AbstractElementType — Type

abstract type AbstractElementType{D, V, E, F, I, P, Q}

Base type for the package that all element types are subtyped off of. The parameters have the following general meaning.

$D$ - Dimension

$V$ - Number of vertices

$E$ - Number of edges

$F$ - Number of faces

$I$ - Interpolation type

$P$ - Polynomial degree

$Q$ - quadrature degree

There are some useful methods associated with this abstract type such as

ReferenceFiniteElements.dimension — Function

dimension(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

Returns dimension D.

ReferenceFiniteElements.num_vertices — Function

num_vertices(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

ReferenceFiniteElements.num_edges — Function

num_edges(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

ReferenceFiniteElements.num_faces — Function

num_faces(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

ReferenceFiniteElements.interpolation_type — Function

interpolation_type(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

Returns the interpolation type $I$.

ReferenceFiniteElements.polynomial_degree — Function

polynomial_degree(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

Returns the interpolation polynomial degree $P$.

polynomial_degree(
    _::Type{<:ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}}
) -> Any

Returns the interpolation polynomial degree $P$.

ReferenceFiniteElements.quadrature_degree — Function

quadrature_degree(
    _::ReferenceFiniteElements.AbstractElementType{D, V, E, F, I, P, Q}
) -> Any

Return the quadrature degree $Q$

That can fetch the parameteric types.

Other useful methods associated with this abstract type are the following which can be used to query information about the element topology.

ReferenceFiniteElements.num_interior_vertices — Function

ReferenceFiniteElements.num_quadrature_points — Function

ReferenceFiniteElements.num_shape_functions — Function

num_shape_functions(
    e::ReferenceFiniteElements.AbstractElementType{D, V, E, F, Lagrange, P, Q}
) -> Any

ReferenceFiniteElements.num_vertices_per_edge — Function

ReferenceFiniteElements.num_vertices_per_face — Function

Element subtypes

Below are additional abstract types subtyped off of AbstractElementType for element topologies of different dimensions and further subtyped based on common element topologies.

0-Dimensional Element types

Different types of 0-D elements (e.g. vertices of various implementations) can be implemented by subtyping off of the abstract type

ReferenceFiniteElements.AbstractVertex — Type

abstract type AbstractVertex{I, P, Q} <: ReferenceFiniteElements.AbstractElementType{0, 1, 0, 0, I, P, Q}

1-Dimensional Element types

Different types of 1-D elements (e.g. edges, sides, lines, etc. of various implementations) can be implemented by subtyping off of the abstract type

ReferenceFiniteElements.AbstractEdge — Type

abstract type AbstractEdge{V, I, P, Q} <: ReferenceFiniteElements.AbstractElementType{1, V, 1, 0, I, P, Q}

2-Dimensional Element types

Different types of 2-D elements (e.g. faces, triangles, quads, polygons, etc. of various implementations) can be implemented by subtyping off of the abstract type

ReferenceFiniteElements.AbstractFace — Type

abstract type AbstractFace{V, E, I, P, Q} <: ReferenceFiniteElements.AbstractElementType{2, V, E, 1, I, P, Q}

There are also various subtypes of this type including

ReferenceFiniteElements.AbstractQuad — Type

abstract type AbstractQuad{V, I, P, Q} <: ReferenceFiniteElements.AbstractFace{V, 4, I, P, Q}

ReferenceFiniteElements.AbstractTri — Type

abstract type AbstractTri{V, I, P, Q} <: ReferenceFiniteElements.AbstractFace{V, 3, I, P, Q}

3-Dimensional Element types

Different types of 3-D elements (e.g. volumes, hexes, tets, etc. of various implementations) can be implemented by subtyping off of the abstract type

ReferenceFiniteElements.AbstractVolume — Type

abstract type AbstractVolume{V, E, F, I, P, Q} <: ReferenceFiniteElements.AbstractElementType{3, V, E, F, I, P, Q}

There are also various subtypes of this type including

ReferenceFiniteElements.AbstractHex — Type

abstract type AbstractHex{V, I, P, Q} <: ReferenceFiniteElements.AbstractVolume{V, 12, 6, I, P, Q}

ReferenceFiniteElements.AbstractTet — Type

abstract type AbstractTet{V, I, P, Q} <: ReferenceFiniteElements.AbstractVolume{V, 6, 4, I, P, Q}

Abstract Types for Storage Containers

ReferenceFiniteElements.AbstractInterpolationType — Type

abstract type AbstractInterpolationType

ReferenceFiniteElements.AbstractInterpolantsContainer — Type

abstract type AbstractInterpolantsContainer{I}