Element types

Below is the currently implemented set of elements organized by vertices, edges, faces, and volumes respectively. If you would like to see an additional element type supported, please file and issue or open a PR.

0-Dimensional Elements

ReferenceFiniteElements.Vertex — Type

struct Vertex{I, P, Q} <: ReferenceFiniteElements.AbstractVertex{I, P, Q}

1-Dimensional Elements

ReferenceFiniteElements.Edge — Type

struct Edge{V, I, P, Q} <: ReferenceFiniteElements.AbstractEdge{V, I, P, Q}

ReferenceFiniteElements.Edge0 — Type

struct Edge0{I, Q} <: ReferenceFiniteElements.AbstractEdge{2, I, 0, Q}

ReferenceFiniteElements.Edge2 — Type

struct Edge2{I, Q} <: ReferenceFiniteElements.AbstractEdge{2, I, 1, Q}

ReferenceFiniteElements.Edge3 — Type

struct Edge3{I, Q} <: ReferenceFiniteElements.AbstractEdge{3, I, 2, Q}

2-Dimensional Elements

ReferenceFiniteElements.Quad — Type

struct Quad{V, I, P, Q} <: ReferenceFiniteElements.AbstractQuad{V, I, P, Q}

ReferenceFiniteElements.Quad0 — Type

struct Quad0{I, Q} <: ReferenceFiniteElements.AbstractQuad{4, I, 0, Q}

ReferenceFiniteElements.Quad4 — Type

struct Quad4{I, Q} <: ReferenceFiniteElements.AbstractQuad{4, I, 1, Q}

ReferenceFiniteElements.Quad9 — Type

struct Quad9{I, Q} <: ReferenceFiniteElements.AbstractQuad{9, I, 2, Q}

ReferenceFiniteElements.Tri — Type

struct Tri{V, I, P, Q} <: ReferenceFiniteElements.AbstractTri{V, I, P, Q}

ReferenceFiniteElements.Tri0 — Type

struct Tri0{I, Q} <: ReferenceFiniteElements.AbstractTri{3, I, 0, Q}

ReferenceFiniteElements.Tri3 — Type

struct Tri3{I, Q} <: ReferenceFiniteElements.AbstractTri{3, I, 1, Q}

ReferenceFiniteElements.Tri6 — Type

struct Tri6{I, Q} <: ReferenceFiniteElements.AbstractTri{6, I, 2, Q}

3-Dimensional Elements

ReferenceFiniteElements.Hex0 — Type

struct Hex0{I, Q} <: ReferenceFiniteElements.AbstractHex{8, I, 0, Q}

ReferenceFiniteElements.Hex8 — Type

struct Hex8{I, Q} <: ReferenceFiniteElements.AbstractHex{8, I, 1, Q}

ReferenceFiniteElements.Tet0 — Type

struct Tet0{I, Q} <: ReferenceFiniteElements.AbstractTet{4, I, 0, Q}

ReferenceFiniteElements.Tet4 — Type

struct Tet4{I, Q} <: ReferenceFiniteElements.AbstractTet{4, I, 1, Q}

ReferenceFiniteElements.Tet10 — Type

struct Tet10{I, Q} <: ReferenceFiniteElements.AbstractTet{10, I, 2, Q}